chore: separate form stories
